% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-extrusion-layer.R, R/layers_fill.R
\name{add_fill_extrusion_layer}
\alias{add_fill_extrusion_layer}
\title{Add a fill extrusion layer to the map}
\usage{
add_fill_extrusion_layer(
  map,
  source = NULL,
  source_layer = NULL,
  filter = NULL,
  fill_extrusion_base = NULL,
  fill_extrusion_color = NULL,
  fill_extrusion_height = NULL,
  fill_extrusion_opacity = NULL,
  fill_extrusion_pattern = NULL,
  fill_extrusion_translate = NULL,
  fill_extrusion_translate_anchor = NULL,
  fill_extrusion_vertical_gradient = TRUE,
  visibility = TRUE,
  popup = NULL,
  id = "fill-extrusion-layer",
  minzoom = NULL,
  maxzoom = NULL
)

add_fill_extrusion_layer(
  map,
  source = NULL,
  source_layer = NULL,
  filter = NULL,
  fill_extrusion_base = NULL,
  fill_extrusion_color = NULL,
  fill_extrusion_height = NULL,
  fill_extrusion_opacity = NULL,
  fill_extrusion_pattern = NULL,
  fill_extrusion_translate = NULL,
  fill_extrusion_translate_anchor = NULL,
  fill_extrusion_vertical_gradient = TRUE,
  visibility = TRUE,
  popup = NULL,
  id = "fill-extrusion-layer",
  minzoom = NULL,
  maxzoom = NULL
)
}
\arguments{
\item{map}{The mapboxer map}

\item{source}{A Mapbox source. Uses the source from the \link{mapboxer} object if no source is supplied.}

\item{source_layer}{The layer to use from a vector tile source (not needed if the source is not vector).}

\item{filter}{A filter expression that is applied to the \code{source}.}

\item{fill_extrusion_base}{The base height of the layer; if \code{NULL}, will default to 0.}

\item{fill_extrusion_color}{(paint) The color of the filled part of this layer.
This color can be specified as rgba with an alpha component and
the color's opacity will not affect the opacity of the 1px stroke, if it is used.}

\item{fill_extrusion_height}{(paint) The height of the fill-extrusion layer.}

\item{fill_extrusion_opacity}{(paint) The opacity of the fill-extrusion layer.}

\item{fill_extrusion_pattern}{(paint) Name of image in sprite to use for drawing image fills.}

\item{fill_extrusion_translate}{(paint) The geometry's offset.
Values are [x, y] where negatives indicate left and up, respectively.}

\item{fill_extrusion_translate_anchor}{(paint) Controls the frame of reference for \code{fill_translate}.
One of "map", "viewport".}

\item{fill_extrusion_vertical_gradient}{(paint) If \code{TRUE} (the default), applies a vertical gradient to the sides of the layer.}

\item{visibility}{(layout) Whether the layer should be displayed.}

\item{popup}{a \link{mapbox_popup} object or a template string.}

\item{id}{The unique id of the layer.}

\item{minzoom}{The minimum zoom at which the layer appears (optional)}

\item{maxzoom}{The maximum zoom at which the layer appears (optional)}
}
\description{
Add a fill extrusion layer to the map

Add a fill-extrusion layer to the map
}
\examples{
LAYER_ID <- "melbourne"

map <- mapboxer(
  center = c(144.9624, -37.8105),
  zoom = 11,
  pitch = 35
) \%>\%
  add_fill_extrusion_layer(
    source = as_mapbox_source(geojsonsf::geo_melbourne),
    fill_extrusion_color = c("get", "fillColor"),
    fill_extrusion_height = list("*", c("get", "AREASQKM"), 1000),
    popup = "Area: {{AREASQKM}} km<sup>2</sup>",
    id = LAYER_ID
  ) \%>\%
  add_filter_control(
    layer_id = LAYER_ID,
    filter = list("<", "AREASQKM", 2),
    rows = 2
  )

if (interactive()) map
 map <- as_mapbox_source(geojsonsf::geo_melbourne) \%>\%
  mapboxer() \%>\%
  set_view_state(
    lng =  144.9624,
    lat = -37.8105,
    zoom = 10,
    pitch = 35
  ) \%>\%
  add_fill_layer(
    fill_color = c("get", "fillColor"),
    fill_opacity = 0.6,
    popup = "Area: {{AREASQKM}} km<sup>2</sup>",
    # AREASQKM > 5
    filter = list(">", c("get", "AREASQKM"), 5)
  )

if (interactive()) map
}
\seealso{
\url{https://docs.mapbox.com/mapbox-gl-js/style-spec/layers/#fill-extrusion}

\url{https://docs.mapbox.com/mapbox-gl-js/style-spec/layers/#fill}
}
