% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-controls.R
\name{add_filter_control}
\alias{add_filter_control}
\title{Add a filter control to the map}
\usage{
add_filter_control(map, layer, filter = NULL, pos = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{layer}{The name/ID of the layer to which the filter is attached.}

\item{filter}{The initial filter expression.}

\item{pos}{The position of the control, one of \code{top-left}, \code{top-right},
\code{bottom-right} or \code{bottom-left}.}
}
\description{
Add a filter control to the map
}
