% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_layer}
\alias{add_layer}
\title{Add a layer to the map}
\usage{
add_layer(map, style, popup = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{style}{A named list that defines the style of the layer.
See \url{https://docs.mapbox.com/mapbox-gl-js/style-spec/layers/} for available style options
for the used layer type.}

\item{popup}{A \href{https://github.com/janl/mustache.js}{mustache} template
in which the tags refer to the properties of the layer's data object.}
}
\description{
Adds any kind of layer to the map.
The type of the layer is specified by the \code{type} property of the
layer definintion.
}
\seealso{
\link{add_popups} for an example of a mustache template used to generate the popup text.
}
