% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{add_marker}
\alias{add_marker}
\title{Add a single marker to the map}
\usage{
add_marker(map, lat, lng, popup = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{lat}{The latitude of the marker.}

\item{lng}{The longitude of the marker.}

\item{popup}{The popup text (HTML) that is displayed when you click on the marker.}
}
\description{
Add a single marker to the map
}
