% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{add_navigation_control}
\alias{add_navigation_control}
\title{Add a navigation control to the map}
\usage{
add_navigation_control(map, ..., pos = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{...}{The properties of the control.}

\item{pos}{The position of the control, one of \code{top-left}, \code{top-right},
\code{bottom-right} or \code{bottom-left}.}
}
\description{
Add a navigation control to the map
}
