% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{add_source}
\alias{add_source}
\alias{add_source.default}
\alias{add_source.json}
\alias{add_source.sf}
\title{Add a data source to the map}
\usage{
add_source(map, data, id, ...)

\method{add_source}{default}(map, data, id, ...)

\method{add_source}{json}(map, data, id, ...)

\method{add_source}{sf}(map, data, id, ...)
}
\arguments{
\item{map}{The map widget.}

\item{data}{The data ...}

\item{id}{The unique id of the data source.}

\item{...}{Optional parameters, see ...}
}
\description{
Add a data source to the map
}
