% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-controls.R
\name{add_text_control}
\alias{add_text_control}
\title{Add a text control to the map}
\usage{
add_text_control(map, text, cssText = NULL, pos = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{text}{The text (HTML) that is displayed.}

\item{cssText}{The \code{cssText} property of the control element.}

\item{pos}{The position of the control, one of \code{top-left}, \code{top-right},
\code{bottom-right} or \code{bottom-left}.}
}
\description{
Add a text control to the map
}
