% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_tooltips}
\alias{add_tooltips}
\title{Add tooltips to a layer}
\usage{
add_tooltips(map, layer_id, tooltip)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{layer_id}{The ID of the layer to which you want to add the tooltips.}

\item{tooltip}{A \href{https://github.com/janl/mustache.js}{mustache} template
in which the tags refer to the properties of the layer's data object.}
}
\description{
Add tooltips to a layer
}
