% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{basemap_background_style}
\alias{basemap_background_style}
\title{Create a background style}
\usage{
basemap_background_style(color = "#111", opacity = 1)
}
\arguments{
\item{color}{The color of the background.}

\item{opacity}{The opacity of the background.}
}
\description{
Create a background style
}
