% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{basemap_carto_style}
\alias{basemap_carto_style}
\title{Use Carto vector tiles as map style}
\usage{
basemap_carto_style(theme = "dark-matter")
}
\arguments{
\item{theme}{The theme of the style, \code{dark-matter}, \code{positron} or \code{voyager}.}
}
\description{
Use Carto vector tiles as map style
}
