% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{basemap_raster_style}
\alias{basemap_raster_style}
\title{Create a raster style}
\usage{
basemap_raster_style(tiles = get_osm_raster_tiles(),
  attribution = NULL)
}
\arguments{
\item{tiles}{A list of tile urls.}

\item{attribution}{The attribution text of the tile layer.}
}
\description{
Create a raster style
}
