% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{create_source}
\alias{create_source}
\alias{create_source.default}
\alias{create_source.json}
\alias{create_source.data.table}
\alias{create_source.sf}
\title{Create a Mapbox data source}
\usage{
create_source(...)

\method{create_source}{default}(...)

\method{create_source}{json}(data, ...)

\method{create_source}{data.table}(data, ..., lng = "lng", lat = "lat")

\method{create_source}{sf}(data, ...)
}
\arguments{
\item{...}{see \href{https://docs.mapbox.com/mapbox-gl-js/style-spec/sources}}

\item{data}{A data frame containing data that can be converted to \code{geojson}.}

\item{lng}{The longitude ...}

\item{lat}{The Latitude ...}
}
\description{
Create a Mapbox data source
}
