% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly-to.R
\name{fly_to}
\alias{fly_to}
\title{Fly to a specific location}
\usage{
fly_to(map, center, ...)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{center}{The location coordinates as a vector in [lat, lng] order}

\item{...}{Optional arguments, see \url{https://docs.mapbox.com/mapbox-gl-js/api/map/#map#flyto}.}
}
\description{
Fly to a specific location
}
