% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap-styles.R
\name{get_basemaps}
\alias{get_basemaps}
\title{Get list of basemaps}
\usage{
get_basemaps()
}
\description{
Get list of basemaps
}
\examples{
basemaps <- get_basemaps()

map <- mapboxer(
  style = basemaps$Carto$positron,
  center = c(-73.9165, 40.7114),
  zoom = 4
)

if (interactive()) map
}
