% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{get_stamen_raster_tiles}
\alias{get_stamen_raster_tiles}
\title{Get Stamen raster tile urls}
\usage{
get_stamen_raster_tiles(theme = "watercolor")
}
\arguments{
\item{theme}{The theme of the tiles.}
}
\description{
Get Stamen raster tile urls
}
