% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapbox.R
\name{mapbox}
\alias{mapbox}
\title{Mapbox Maps from R}
\usage{
mapbox(df, htmlfile = "index.html")
}
\arguments{
\item{df}{Input data.frame. Must contain fields called latitude and longitude}

\item{htmlfile}{Destination file name. Default is \code{index.html}}
}
\description{
This function takes an input data.frame and writes out a html file containing a Mapbox map and all javascript dependencies.
}

