% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popups.R
\name{mapbox_popup}
\alias{mapbox_popup}
\title{Create a popup object}
\usage{
mapbox_popup(template, ..., event = c("click", "hover"))
}
\arguments{
\item{template}{A \href{https://github.com/janl/mustache.js}{mustache} template
in which the tags refer to the properties of the layer's data object.}

\item{...}{The options of the popup element, see see https://docs.mapbox.com/mapbox-gl-js/api/markers/#popup}

\item{event}{The event on which the popup is displayed.}
}
\description{
Create a popup object
}
