% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{mapbox_source}
\alias{mapbox_source}
\title{Create a Mapbox source}
\usage{
mapbox_source(type, ...)
}
\arguments{
\item{type}{The type of the source, e. g. \code{geojson}.}

\item{...}{The properties of the source.
See \url{https://docs.mapbox.com/mapbox-gl-js/style-spec/sources} for available options
for the given source type.}
}
\description{
Create a Mapbox source
}
