% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxer-proxy.R
\name{mapboxer_proxy}
\alias{mapboxer_proxy}
\title{Create a mapboxer proxy object}
\usage{
mapboxer_proxy(shiny_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shiny_id}{The output ID of the mapboxer object that should be updated.}

\item{session}{The current Shiny session object. Usually the default value can be used.}
}
\description{
Create a \code{\link{mapboxer}}-like object that can be used to update a mapboxer object
that has already been rendered in a Shiny app.
}
