% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{mapboxer_use_v2}
\alias{mapboxer_use_v2}
\title{Switch between Mapbox GL JS v1 or v2 in a mapboxer project}
\usage{
mapboxer_use_v2(use_v2)
}
\arguments{
\item{use_v2}{If \code{TRUE}, mapboxer will use Mapbox GL JS version 2, which includes new features (such as Globe view and non-Mercator projections) but is less permissively licensed than Mapbox GL JS v1.  Please review the Mapbox terms of service at https://www.mapbox.com/legal/tos/ for more information.}
}
\value{
Returns the vale of \code{use_v2} invisibly.
}
\description{
Switch between Mapbox GL JS v1 or v2 in a mapboxer project
}
\examples{
\dontrun{

library(mapboxer)
mapboxer_use_v2(TRUE)

# Globe view is only available in V2
mapboxer(
 center = c(-73.9165, 40.7114),
 style = basemaps$Mapbox$streets_v11,
 zoom = 2,
 projection = "globe"
)

}
}
