% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{mapboxer_use_version}
\alias{mapboxer_use_version}
\title{Switch between Mapbox GL JS versions in a mapboxer project}
\usage{
mapboxer_use_version(version = c("mapbox-gl", "mapbox-gl2", "mapbox-gl3"))
}
\arguments{
\item{version}{The version to use}
}
\value{
Returns the value of \code{version} invisibly.
}
\description{
Switch between Mapbox GL JS versions in a mapboxer project
}
\examples{
\dontrun{

library(mapboxer)
mapboxer_use_version("mapbox-gl3")

# Globe view is only available in V2
mapboxer(
 center = c(-73.9165, 40.7114),
 style = basemaps$Mapbox$streets_v11,
 zoom = 2,
 projection = "globe"
)

}
}
