% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motor_vehicle_collisions_nyc}
\alias{motor_vehicle_collisions_nyc}
\title{Motor Vehicle Collisions in NYC}
\format{
A data frame with 1601 rows and 6 variables,
  where each row is a Motor Vehicle Collision:
\describe{
  \item{date}{occurrence date of collision}
  \item{time}{occurrence time of collision}
  \item{lng}{latitude coordinate for Global Coordinate System, WGS 1984, decimal degrees (EPSG 4326)}
  \item{lat}{longitude coordinate for Global Coordinate System, WGS 1984, decimal degrees (EPSG 4326)}
  \item{injured}{number of persons injured}
  \item{killed}{number of persons killed}
}
}
\source{
\url{https://opendata.cityofnewyork.us/}
}
\usage{
motor_vehicle_collisions_nyc
}
\description{
Motor Vehicle Collisions in NYC
}
\keyword{datasets}
