% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{remove_source}
\alias{remove_source}
\title{Remove a Mapbox source from the map}
\usage{
remove_source(map, id = "mapboxer")
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{id}{The unique id of the data source to remove.}
}
\description{
Remove a Mapbox source from the map
}
