% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxer-proxy.R
\name{send_update}
\alias{send_update}
\title{Update a mapboxer object in a Shiny app}
\usage{
send_update(proxy_obj, ...)
}
\arguments{
\item{proxy_obj}{A \code{\link{mapboxer_proxy}} object.}

\item{...}{unused}
}
\description{
Update a mapboxer object in a Shiny app
}
