% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_update.R
\name{set_feature_state}
\alias{set_feature_state}
\title{Set feature state
See https://docs.mapbox.com/mapbox-gl-js/api/map/#map#setfeaturestate}
\usage{
set_feature_state(
  map,
  feature_id,
  feature_source,
  feature_source_layer = NULL,
  state
)
}
\arguments{
\item{feature_id}{The id of the feature to set the state for (one at a time)}

\item{feature_source}{The source}

\item{feature_source_layer}{The source layer (required for vector layers)}

\item{source}{A Mapbox source.}
}
\description{
Set feature state
See https://docs.mapbox.com/mapbox-gl-js/api/map/#map#setfeaturestate
}
