% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{set_filter}
\alias{set_filter}
\title{Update the filter of a layer}
\usage{
set_filter(map, layer_id, filter)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{layer_id}{The ID of the layer whose property should be updated.}

\item{filter}{A filter expression that is applied to the \code{source}.}
}
\description{
Update the filter of a layer
}
