% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fog_terrain.R
\name{set_fog}
\alias{set_fog}
\title{Set the fog properties for a map}
\usage{
set_fog(map, ...)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{...}{Optional arguments, see \url{https://docs.mapbox.com/mapbox-gl-js/api/map/#map#setfog}.}
}
\description{
Set the fog properties for a map
}
