% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_update.R
\name{set_layer_properties}
\alias{set_layer_properties}
\alias{set_paint_property}
\alias{set_layout_property}
\title{Update layer properties}
\usage{
set_paint_property(map, layer_id, property, value)

set_layout_property(map, layer_id, property, value)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{layer_id}{The ID of the layer whose property should be updated.}

\item{property}{The name of the property that should be updated.}

\item{value}{The new value of the property.}
}
\description{
Update layer properties
}
\section{Functions}{
\itemize{
\item \code{set_paint_property()}: Update a paint property of a layer.

\item \code{set_layout_property()}: Update a layout property of a layer.

}}
