% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{set_map_style}
\alias{set_map_style}
\title{Set the style of the map}
\usage{
set_map_style(map, style)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{style}{The map's Mapbox style.}
}
\description{
Set the style of the map
}
