% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{set_source_tiles}
\alias{set_source_tiles}
\title{Update the tiles URL of a vector source}
\usage{
set_source_tiles(map, source_id, tiles)
}
\arguments{
\item{source_id}{The id of the data source to update}

\item{tiles}{A new list of tile urls}
}
\description{
Update the tiles URL of a vector source
}
