% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fog_terrain.R
\name{set_terrain}
\alias{set_terrain}
\title{Set the terrain properties for a map}
\usage{
set_terrain(map, terrain_source, exaggeration = 1.5, ...)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{terrain_source}{A Mapbox source of type 'raster-dem'}

\item{exaggeration}{The exaggeration factor}

\item{...}{Optional arguments, see \url{https://docs.mapbox.com/mapbox-gl-js/example/add-terrain/}.}
}
\description{
Set the terrain properties for a map
}
