% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-state.R
\name{set_view_state}
\alias{set_view_state}
\title{Set the view state of the map}
\usage{
set_view_state(map, lng, lat, zoom = 9, pitch = 0, bearing = 0)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{lng}{The longitude of the  geographical center point of the map.}

\item{lat}{The latitude of the geographical center point of the map.}

\item{zoom}{The zoom level of the map.}

\item{pitch}{The pitch (tilt) of the map.}

\item{bearing}{The bearing (rotation) of the map.}
}
\description{
Set the view state of the map
}
