% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{stamen_raster_tiles}
\alias{stamen_raster_tiles}
\title{Get Stamen raster tile URLs}
\usage{
stamen_raster_tiles(theme = c("watercolor"))
}
\arguments{
\item{theme}{The theme of the tiles.}
}
\description{
Get Stamen raster tile URLs
}
