% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxer-proxy.R
\name{update_mapboxer}
\alias{update_mapboxer}
\title{Update a mapboxer proxy object in a Shiny app}
\usage{
update_mapboxer(proxy_obj, ...)
}
\arguments{
\item{proxy_obj}{A \code{\link{mapboxer_proxy}} object.}

\item{...}{unused}
}
\description{
Update a mapboxer proxy object in a Shiny app
}
