% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{use_raster_style}
\alias{use_raster_style}
\title{Use raster tiles as map style}
\usage{
use_raster_style(tiles = get_osm_raster_tiles(), attribution = NULL)
}
\arguments{
\item{tiles}{A list of tile urls.}

\item{attribution}{The attribution text of the tile layer.}
}
\description{
Use raster tiles as map style
}
