% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{use_stamen_raster_style}
\alias{use_stamen_raster_style}
\title{Use Stamen raster tiles as map style}
\usage{
use_stamen_raster_style(theme = "watercolor")
}
\arguments{
\item{theme}{The theme of the tiles.}
}
\description{
Use Stamen raster tiles as map style
}
