% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_deck-gl-layer.R
\name{add_deckgl_layer}
\alias{add_deckgl_layer}
\title{Add a deck.gl layer to the map (experimental)}
\usage{
add_deckgl_layer(map, type, data, ..., id = "deckgl-layer",
  popup = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object}

\item{type}{The type of the deck.gl layer, e. g. \code{ScatterplotLayer}.}

\item{data}{The data object.}

\item{...}{The properties of the layer.}

\item{id}{The unique ID of the layer}

\item{popup}{A mustache template string.}
}
\description{
Add a deck.gl layer to the map (experimental)
}
\examples{
# deck.gl HexagonLayer

map <- mapboxer(
  center = c(-73.9165, 40.7114),
  zoom = 10,
  pitch = 35
) \%>\%
  add_deckgl_layer(
    type = "HexagonLayer",
    data = motor_vehicle_collisions_nyc,
    extruded = TRUE,
    radius = 300,
    elevationScale = 4,
    getPosition = "@=[{{lng}}, {{lat}}]",
    opacity = 0.5,
    popup = "Number of crashes: {{points.length}}"
  ) \%>\%
  add_navigation_control()

if (interactive()) map
}
