% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-control.R
\name{add_draw_control}
\alias{add_draw_control}
\title{Add a draw control to the map (experimental)}
\usage{
add_draw_control(map, ..., pos = NULL, data = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{...}{The options of the control. If named, converted into an object passed to the constructor; if not, converted into arguments passed to tne constructor in the same order.}

\item{pos}{The position of the control. One of \code{top-left}, \code{top-right},
\code{bottom-right} or \code{bottom-left}.}

\item{data}{A \code{GeoJSON} or \code{sf} object.}
}
\description{
Add a draw control to the map (experimental)
}
\seealso{
\url{https://github.com/mapbox/mapbox-gl-draw/blob/main/docs/API.md}
  for available options for the draw control.
}
