% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-extrusion-layer.R
\name{add_fill_extrusion_layer}
\alias{add_fill_extrusion_layer}
\title{Add a fill extrusion layer to the map}
\usage{
add_fill_extrusion_layer(map, source = NULL, filter = NULL,
  fill_extrusion_base = NULL, fill_extrusion_color = NULL,
  fill_extrusion_height = NULL, fill_extrusion_opacity = NULL,
  fill_extrusion_pattern = NULL, fill_extrusion_translate = NULL,
  fill_extrusion_translate_anchor = NULL,
  fill_extrusion_vertical_gradient = NULL, visibility = c("visible",
  "none"), popup = NULL, id = "fill-extrusion-layer")
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{source}{A Mapbox source. Uses the source from the \link{mapboxer} object if no source is supplied.}

\item{filter}{A filter expression that is applied to the \code{source}.}

\item{fill_extrusion_base}{(paint) The height with which to extrude the base of this layer.
Must be less than or equal to \code{fill_extrusion_height}.}

\item{fill_extrusion_color}{(paint) The (base) color of the extruded fill.}

\item{fill_extrusion_height}{(paint) The height with which to extrude this layer. Units in meters.}

\item{fill_extrusion_opacity}{(paint) The opacity of the entire fill extrusion layer.
This is rendered on a per-layer, not per-feature, basis, and data-driven styling is not available.}

\item{fill_extrusion_pattern}{(paint) Name of image in sprite to use for drawing images on extruded fills.
For seamless patterns, image width and height must be a factor of two (2, 4, 8, ..., 512).}

\item{fill_extrusion_translate}{(paint) The geometry's offset. Values are [x, y] where negatives indicate left and up (on the flat plane), respectively.
Units in pixels.}

\item{fill_extrusion_translate_anchor}{(paint) Controls the frame of reference for \code{fill_extrusion_translate}.
One of "map", "viewport".}

\item{fill_extrusion_vertical_gradient}{(paint) Whether to apply a vertical gradient to the sides of a fill-extrusion layer.
If true, sides will be shaded slightly darker farther down.}

\item{visibility}{(layout) Whether the layer should be displayed.}

\item{popup}{A \href{https://github.com/janl/mustache.js}{mustache} template
in which the tags refer to the properties of the layer's data object.}

\item{id}{The unique id of the layer.}
}
\description{
Add a fill extrusion layer to the map
}
\examples{
LAYER_ID <- "melbourne"

map <- mapboxer(
  center = c(144.9624, -37.8105),
  zoom = 11,
  pitch = 35
) \%>\%
  add_fill_extrusion_layer(
    source = as_mapbox_source(geojsonsf::geo_melbourne),
    fill_extrusion_color = c("get", "fillColor"),
    fill_extrusion_height = list("*", c("get", "AREASQKM"), 1000),
    popup = "Area: {{AREASQKM}} km<sup>2</sup>",
    id = LAYER_ID
  ) \%>\%
  add_filter_control(
    layer_id = LAYER_ID,
    filter = list("<", "AREASQKM", 2),
    rows = 2
  )

if (interactive()) map
}
\seealso{
\url{https://docs.mapbox.com/mapbox-gl-js/style-spec/layers/#fill-extrusion}
}
