% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-controls.R
\name{add_filter_control}
\alias{add_filter_control}
\title{Add a filter control to the map}
\usage{
add_filter_control(
  map,
  layer_id,
  filter = NULL,
  pos = NULL,
  rows = 1,
  cols = 20
)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{layer_id}{The ID of the layer to which the filter is attached.}

\item{filter}{The initial filter expression.}

\item{pos}{The position of the control. One of \code{top-left}, \code{top-right},
\code{bottom-right} or \code{bottom-left}.}

\item{rows}{The number of rows of the textarea input.}

\item{cols}{The number of columns of the textarea input.}
}
\description{
Add a filter control to the map
}
