% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{add_legend_control}
\alias{add_legend_control}
\title{Add a legend control to the map}
\usage{
add_legend_control(map, colors, labels, title = NULL,
  pos = c("bottom-right", "bottom-left", "top-left", "top-right"))
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{colors}{The colors of the legend items.}

\item{labels}{The labels of the legend items.}

\item{title}{The title of the legend.}

\item{pos}{The position of the control. One of \code{top-left}, \code{top-right},
\code{bottom-right} or \code{bottom-left}.}
}
\description{
Add a legend control to the map
}
