% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{add_popups}
\alias{add_popups}
\title{Add popups for a given layer to the map}
\usage{
add_popups(map, layer, popup)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{layer}{The name of the layer.}

\item{popup}{A \href{https://github.com/janl/mustache.js}{mustache} template string.}
}
\description{
Add popups for a given layer to the map
}
