% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-controls.R
\name{add_text_control}
\alias{add_text_control}
\title{Add a text control to the map}
\usage{
add_text_control(map, text, pos = NULL, css_text = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{text}{The text (HTML) that is displayed.}

\item{pos}{The position of the control, one of \code{top-left}, \code{top-right},
\code{bottom-right} or \code{bottom-left}.}

\item{css_text}{A \code{cssText} string that will modify the style of the control element.}
}
\description{
Add a text control to the map
}
