% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popups.R
\name{add_tooltips}
\alias{add_tooltips}
\title{Add tooltips to a layer (deprecated)}
\usage{
add_tooltips(map, ...)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{...}{ignored}
}
\description{
Add tooltips to a layer (deprecated)
}
