% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{basemap_background_style}
\alias{basemap_background_style}
\title{Use a background color as map style}
\usage{
basemap_background_style(color = "#111")
}
\arguments{
\item{color}{The color of the background.}
}
\description{
Use a background color as map style
}
