% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{basemap_raster_style}
\alias{basemap_raster_style}
\title{Create a raster style}
\usage{
basemap_raster_style(
  tiles = stamen_raster_tiles("terrain"),
  attribution = NULL
)
}
\arguments{
\item{tiles}{A list of tile URLs.}

\item{attribution}{The attribution text of the tile layer.}
}
\description{
Creates a raster style that can be used as a basemap.
}
