% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basemaps}
\alias{basemaps}
\title{A list of basemap style URLs}
\format{
An object of class \code{list} of length 2.
}
\usage{
basemaps
}
\description{
A list of basemap style URLs
}
\keyword{datasets}
