% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxer.R
\name{mapboxer}
\alias{mapboxer}
\title{Create a mapboxer widget}
\usage{
mapboxer(source = NULL, style = use_carto_style(), ..., width = NULL,
  height = NULL, element_id = NULL)
}
\arguments{
\item{source}{...}

\item{style}{The map's Mapbox style.}

\item{...}{Map properties ...}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{element_id}{...}
}
\description{
Create a mapboxer widget
}
