% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_update.R
\name{set_data}
\alias{set_data}
\title{Update the data of a Mapbox source}
\usage{
set_data(map, data, source_id = NULL)
}
\arguments{
\item{map}{A \link{mapboxer} object.}

\item{data}{A data frame that contains longitudes and latitudes in separate columns
or an \code{sf}-object.}

\item{source_id}{The ID of the source whose data should be updated.}
}
\description{
Update the data of a Mapbox source
}
